+++ documentation/anderes/changelog.md
@@ -16,7 +16,9 @@ icon: clock-rotate-left
 ### 🔧 Änderungen
 
 * Parameter `serverdomain` von `Vertretungsplan` in `domain` umbenannt
-* Docstring zu `Stunde.fach` etwas konkretisiert
+* Methode `stundenInPeriode()` von `Klasse`, `Lehrer` und `Raum` entfernt
+* `Stunde.lehrergeändert`, `~.raumgeändert` und `.~klassegeändert` können bei respektiven Plantypen nun nicht mehr `None` sein. Stattdessen wird `~.geändert` weitergegeben
+* Alle Docstrings wurden überarbeitet
 
 ## [2.1.0.1](https://pypi.org/project/vpmobil/2.1.0.1/) (2026-03-27)